cc/cloud-function/function-1/requirements.txt
functions-framework
Pillow
numpy